\name{BreakoutDetection-package}
\alias{BreakoutDetection-package}
\alias{BreakoutDetection}
\docType{package}
\title{
	Breakout Detector: Energy Divisive with Medians
}
\description{
This package performs single and multiple change point analysis of univariate time series.
The procedures in this package are designed to be robust to the existence of anomalies/outliers 
within the time series.
}
\details{
\tabular{ll}{
Package: \tab BreakoutDetection\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2014-10-07\cr
License: \tab GPL(>= 2)\cr
}
}
\author{
Nicholas A. James, Arun Kejariwal, David S. Matteson

Maintainer: Nicholas A. James <nj89@cornell.edu>
}
\references{
Nicholas A. James, Arun Kejariwal, David S. Matteson, "Leveraging Cloud Data to Mitigate User Experience from 'Breaking Bad': The Twitter Approach, 2014
}
